ibr: DO NOT COMMIT: make the example work
This will have the example create an account via IBR, bind on the same
stream and then disconnect.

The hacky bits here are about event handling, and also about how XMPP
features are handled.

At the moment, we need to register the IBR plugin before we register
stream features, otherwise slixmpp will attempt to bind before even
considering IBR.

We also have to make slixmpp wait that the registration is finished and
not hand over the processing to other handlers, which causes it to fail
horribly (attempt bind while doing registration, etc.).

Signed-off-by: Maxime “pep” Buquet <pep@bouah.net>