xep_0384: when encrypting, only tell client when trust is undecided
On UntrustedException, there are two possibilities.  Either trust has not been
explicitely set yet, and is 'undecided', or the device is explicitely not
trusted. When undecided, we need to ask our user to make a choice. If
untrusted, then we can safely tell the OMEMO lib to not encrypt to this device.

Signed-off-by: Maxime “pep” Buquet <pep@bouah.net>